+++ csw/SequencerCommandService.py
@@ -7,11 +7,11 @@ class SequencerCommandService:
     async def submit(self, sequence: Sequence) -> SubmitResponse:
         pass
 
-    async def submitAndWait(self, sequence: Sequence, timeoutInSeconds: int) -> SubmitResponse:
+    async def submitAndWait(self, sequence: Sequence, timeoutInSeconds: float) -> SubmitResponse:
         pass
 
     async def query(self, runId: str) -> SubmitResponse:
         pass
 
-    async def queryFinal(self, runId: str, timeoutInSeconds: int = 10) -> SubmitResponse:
+    async def queryFinal(self, runId: str, timeoutInSeconds: float = 10) -> SubmitResponse:
         pass
